Refactors Scoped to make serializable_hash more controllable via super
