+++ features/0002-links_changes.md
@@ -25,7 +25,8 @@ SetMember should have an Relation class that inherits from it.  Relations logic 
 - [x] Move serialization logic out of [scope][scope].
 - [x] LinkSets should be merge-able.
 - [x] **Links::Self < Links::SetMember**. link should be part of the link_set, and it's serialization should be managed by link set.  Currently, it's just modifying the output json.
-- [ ] Scoped.rb inheritance refactoring to remove the merge_links, merge_inclusions calls.
+- [x] Scoped.rb inheritance refactoring to remove the merge_links, merge_inclusions calls.
+- [ ] `merge_inclusions!` should use links config.
 - [ ] refactor the internal app.
 - [ ] Remove The InclusionSupport class dependency in the models and controllers, just leaving link support.
 - [ ] set_member associations has a `find` method.
